修改提交内容

+++ include/Sea.hpp
@@ -96,27 +96,27 @@ extern "C++"
     // 只提供syscall内存读取的方法 过CRC校验请自行绕道 暂且提供D/F类型的数据更改和读取
 
     int initPackage(std::string packageName, bool printLog);     // 初始化内存广角
-    bool mem_vm_readv(long address, void *buffer, size_t size);  // 公开接口 可以实现结构体的读取
-    bool mem_vm_writev(long address, void *buffer, size_t size); // 公开接口 同上
-    long Mem_lsp64(long addres);                                 // 指针跳转 没有32位 谁爱用谁写
+    bool mem_vm_readv(uintptr_t address, void *buffer, size_t size);  // 公开接口 可以实现结构体的读取
+    bool mem_vm_writev(uintptr_t address, void *buffer, size_t size); // 公开接口 同上
+    uintptr_t Mem_lsp64(uintptr_t addres);                                 // 指针跳转 没有32位 谁爱用谁写
     // 数据声明 原本是想要用aoto封装读写 但是考虑到小小白也不会 就舍弃了
-    float Mem_getFloat(long addres);             // 获取地址的Float
-    int Mem_getDword(long addres);               // 获取内存的Dworld
-    void Mem_WriteFloat(long addres, float fix); // 写入指定内存的Float
-    void Mem_WriteDword(long addres, int fix);   // 写入指定内存的Dworld
+    float Mem_getFloat(uintptr_t addres);             // 获取地址的Float
+    int Mem_getDword(uintptr_t addres);               // 获取内存的Dworld
+    void Mem_WriteFloat(uintptr_t addres, float fix); // 写入指定内存的Float
+    void Mem_WriteDword(uintptr_t addres, int fix);   // 写入指定内存的Dworld
 
     // 获取模块首地址
-    long Mem_get_module_cb(const char *name, int index);
-    long Mem_get_module_cb(const char *name, int index, bool isSplit, const char *txt); // 具体就加了一个条件 因为有些内存字段在r-rw 有些在r-p
-    long Mem_get_module(const char *name, int index);
-    long Mem_get_module(char *module_name); // 内核模块获取方式 注意区分
+    uintptr_t Mem_get_module_cb(const char *name, int index);
+    uintptr_t Mem_get_module_cb(const char *name, int index, bool isSplit, const char *txt); // 具体就加了一个条件 因为有些内存字段在r-rw 有些在r-p
+    uintptr_t Mem_get_module(const char *name, int index);
+    uintptr_t Mem_get_module(char *module_name); // 内核模块获取方式 注意区分
 
     //-------------------内存管理拓展
-    long Mem_readPointer(long addr, long *arr, int sz);       // 读取数组链条 传入数组
-    long Mem_readPointer(long orgin, std::vector<long> &vec); // 读取指针 升级版噢
+    uintptr_t Mem_readPointer(uintptr_t addr, uintptr_t *arr, int sz);       // 读取数组链条 传入数组
+    uintptr_t Mem_readPointer(uintptr_t orgin, std::vector<uintptr_t> &vec); // 读取指针 升级版噢
 
-    // QX驱动模块 目前支持QX8.2
-    bool initQX(bool info); // 开启QX内核支持 当然框架肯定就废了 自行抉择
+    // 驱动模块 目前支持 AC 驱动 哈哈哈哈 自用 5.10 5.15 4.14.168 其他的我不管 因为自用
+    bool initQX(bool info); // 开启内核支持 当然框架肯定就废了 自行抉择
 
 #ifdef __cplusplus
 }
